+++ tests/client_gen/example_program_gen/types/__init__.py
@@ -5,3 +5,5 @@ from . import foo_struct
 from .foo_struct import FooStruct, FooStructJSON
 from . import foo_enum
 from .foo_enum import FooEnumKind, FooEnumJSON
+from . import u8_array
+from .u8_array import U8Array