+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.19.0] - 2024-02-12
+
+### Added
+
+- Add support for alias types in IDL [#(129)](https://github.com/kevinheavey/anchorpy/pull/129)
+
+### Changed
+
+- Make the Pytest plugin dependencies optional. Install with `pip install anchorpy[pytest]` if you 
+want the Pytest plugin [#(129)](https://github.com/kevinheavey/anchorpy/pull/129)
+- Update to Anchor 0.29.0 [#(132)](https://github.com/kevinheavey/anchorpy/pull/132)
+- Update to solders 0.20.0 [#(132)](https://github.com/kevinheavey/anchorpy/pull/132)
+- Update to solana-py 0.32.0 [#(132)](https://github.com/kevinheavey/anchorpy/pull/132)
+
 ## [0.18.0] - 2023-07-20
 
 ### Changed